Add files via upload
The images taken from all four datasets individually are grouped into a few cohesive clusters, making this study an unsupervised problem. The ultimate aim of contactless authentication for vein pattern recognition is to anticipate the clusters that correspond to the data and labels specified in each data point. The performance of this algorithm can be defined by the correct identification rate (CIR). This can be computed from the Receiver Operating Characteristic (ROC) curve as the area under the curve (AUC).